+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to your `assets/backend/package.json`:
 * Remove datasource block and introduce module blocks with datasource instead.
   Run `mix brando.upgrade && mix ecto.migrate` to convert your existing datasource
   blocks to module blocks.
+* Add scheduled publishing for revisions
 * Fix max width for #content
 * Automatically add uploaded gallery images to gallery
 * Add `alert` and `after_save` to forms: